add awww wallpaper as future

+++ .local/bin/chBackGround.sh
@@ -1,24 +1,20 @@
 #! /usr/bin/env bash
-if [[ -a "$(dirname "$0")/clientid.sh" ]]; then
-	source "$(dirname "$0")/clientid.sh"
-else
-	echo "clientid file not find"
-fi
-
-COUNT=30
 if [[ -a ~/Pictures/1.jpg ]]; then
-	mv -f ~/Pictures/1.jpg ~/Pictures/wallpaper.jpg
+	echo "mv -f ~/Pictures/1.jpg ~/Pictures/wallpaper.jpg"
+	cp -f ~/Pictures/1.jpg ~/Pictures/wallpaper.jpg
 fi
-pkill swaybg 2>/dev/null || true
 
-swaybg -i ~/Pictures/wallpaper.jpg -m fill & 
+if command -v awww >/dev/null ; then
+awww img "$HOME/Pictures/wallpaper.jpg" \
+  --transition-type random \
+  --transition-duration 2 \
+  --transition-fps 60
+#awww img ~/Pictures/wallpaper.jpg
+else
 
-Json=$(curl -H "Authorization: Client-ID $UNSPLASH_CLIENT_ID" \
-	"https://api.unsplash.com/photos/random?orientation=landscape&count=$COUNT")
+	pkill swaybg 2>/dev/null || true
+	swaybg -i ~/Pictures/wallpaper.jpg -m fill & 
+fi
 
-
-mapfile -t IMG_URLS < <(echo "$Json" | jq -r 'if type=="array" then .[].urls.raw else .urls.raw end')
-
-for i in "${!IMG_URLS[@]}"; do
-	curl -L "${IMG_URLS[$i]}" -o "$HOME/Pictures/$((i + 1)).jpg"
-done
+pkill -f "/home/bi/.local/bin/downloadPictures.sh"
+bash -c "$(dirname "$0")/downloadPictures.sh"
